[IMP] account_multicompany_code: fixes pending from migration
- Split the models in individual files
- Add the code field to tree and kanban view
- Define a setUpClass method for tests
- Set a SQL constraint "code_unique" for efficiency
- Rename file: company_view -> res_company_views.xml
