+++ account_multicompany_code/i18n/es.po
@@ -6,9 +6,9 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 15.0+e\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-10-06 03:52+0000\n"
-"PO-Revision-Date: 2022-10-19 18:41+0000\n"
-"Last-Translator: Luis González <lgonzalez@vauxoo.com>\n"
+"POT-Creation-Date: 2022-11-04 18:37+0000\n"
+"PO-Revision-Date: 2022-11-04 18:37+0000\n"
+"Last-Translator: Luis Manuel <luis.manuel@vauxoo.com>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -26,8 +26,7 @@ msgid "Code"
 msgstr "Código"
 
 #. module: account_multicompany_code
-#: code:addons/account_multicompany_code/models/res_company.py:0
-#, python-format
+#: model:ir.model.constraint,message:account_multicompany_code.constraint_res_company_unique_code
 msgid "Code must be unique"
 msgstr "El código debe ser único"
 
@@ -35,20 +34,6 @@ msgstr "El código debe ser único"
 #: model:ir.model,name:account_multicompany_code.model_res_company
 msgid "Companies"
 msgstr "Compañías"
-
-#. module: account_multicompany_code
-#: model:ir.model.fields,field_description:account_multicompany_code.field_account_account__display_name
-#: model:ir.model.fields,field_description:account_multicompany_code.field_account_journal__display_name
-#: model:ir.model.fields,field_description:account_multicompany_code.field_res_company__display_name
-msgid "Display Name"
-msgstr "Nombre mostrado"
-
-#. module: account_multicompany_code
-#: model:ir.model.fields,field_description:account_multicompany_code.field_account_account__id
-#: model:ir.model.fields,field_description:account_multicompany_code.field_account_journal__id
-#: model:ir.model.fields,field_description:account_multicompany_code.field_res_company__id
-msgid "ID"
-msgstr ""
 
 #. module: account_multicompany_code
 #: model:ir.model.fields,help:account_multicompany_code.field_res_company__code
@@ -59,10 +44,3 @@ msgstr "Código interno de la compañía"
 #: model:ir.model,name:account_multicompany_code.model_account_journal
 msgid "Journal"
 msgstr "Diario"
-
-#. module: account_multicompany_code
-#: model:ir.model.fields,field_description:account_multicompany_code.field_account_account____last_update
-#: model:ir.model.fields,field_description:account_multicompany_code.field_account_journal____last_update
-#: model:ir.model.fields,field_description:account_multicompany_code.field_res_company____last_update
-msgid "Last Modified on"
-msgstr "Última modificación el"
