+++ account_multicompany_code/models/__init__.py
@@ -1,2 +1,3 @@
+from . import account_account
+from . import account_journal
 from . import res_company
-from . import account
