--- account_multicompany_code/views/company_view.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <record id="view_company_form" model="ir.ui.view">
-        <field name="name">qf_common.company.form</field>
-        <field name="model">res.company</field>
-        <field name="inherit_id" ref="base.view_company_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='partner_id']" position="before">
-                <field name="code" required="1" />
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
